segmentation/requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pillow==10.0.0
matplotlib==3.7.2
tqdm==4.66.1
scipy==1.11.2
scikit-learn==1.3.0
opencv-python==4.8.0.76